Moving prescriptions AlertBox component to src/js/common.
This component will also be used as part of Secure Messaging. Making it a common component.
