+++ src/js/common/components/AlertBox.jsx
